klientctl: move sync subcommand to mount namespace
